svg/text/inheritance.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="8000"><title>Inheritance of text properties</title><metadata><h:link href="https://svgwg.org/svg2-draft/text.html" rel="help"/><h:meta content="text properties do not inherit." name="assert"/><h:meta content="text properties have expected initial values." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/inheritance-testcommon.js"/><script><![CDATA[

assert_not_inherited('inline-size', 'auto', '10px');
assert_not_inherited('shape-inside', 'auto', 'url("https://example/com/")');
assert_not_inherited('shape-subtract', 'auto', 'url("https://example/com/")');
assert_not_inherited('shape-margin', '0px', '10px');
assert_inherited('text-anchor', 'start', 'middle');
assert_not_inherited('text-decoration-fill', 'none', 'url("https://example/com/")');
assert_not_inherited('text-decoration-stroke', 'none', 'url("https://example/com/")');

  ]]></script></svg>
