third_party/blink/web_tests/external/wpt/svg/text/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property shape-inside has initial value auto
  assert_true: shape-inside doesn't seem to be supported in the computed style expected true got false
[FAIL] Property shape-inside does not inherit
  assert_true: expected true got false
[FAIL] Property shape-subtract has initial value auto
  assert_true: shape-subtract doesn't seem to be supported in the computed style expected true got false
[FAIL] Property shape-subtract does not inherit
  assert_true: expected true got false
[FAIL] Property text-decoration-fill has initial value none
  assert_true: text-decoration-fill doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-fill does not inherit
  assert_true: expected true got false
[FAIL] Property text-decoration-stroke has initial value none
  assert_true: text-decoration-stroke doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-stroke does not inherit
  assert_true: expected true got false
Harness: the test ran to completion.
